Improve the logging report
